fix(reader): 他藏对读 改为内联侧栏，不再遮挡经文
用户实测反馈：antd Drawer 作 560px overlay 覆盖了右侧经文。虽然左侧
经文仍可见，但阅读区右半部被压占，与 ReaderAIPanel 同为右侧面板时
更显突兀。

改为**内联 flex 侧栏**，和 ReaderAIPanel 同一模式：

- 新组件 ReaderParallelPanel（无 Drawer 容器，纯内容）
- TextReaderPage 里作为 reader-with-sidebar 的 flex 子元素
- 独立 reader-ai-divider 拖拽条支持调宽（320-720px，默认 480）
- 独立关闭按钮；他藏对读按钮加 active 态（primary 蓝）
- 他藏对读面板在 AI 面板左侧，两者可同时开启，各自独立拖拽

打开时行为：
- 主 reader-container 用 flex:1 自动收缩
- 两个面板 flex-shrink:0 固定宽度
- 经文永不被遮挡，横向 fit viewport

删除：ReaderParallelDrawer.tsx（已由 ReaderParallelPanel 替代）
保留：所有 API 调用、Collapse 内容渲染、字体 fallback 逻辑

CSS 改动：
- .reader-parallel-panel: 复用 AI panel 的 flex column + overflow 容器模式
- reader-with-sidebar 的 reader-ai-active 类现在在 aiPanelOpen 或
  parallelPanelOpen 任一为 true 时激活（锁高度 + 独立滚动）
